requirements.txt
Flask==2.2.5
openai==1.13.3
gitpython==3.1.32
PyGithub==2.2.0
PyYAML==6.0.1
Requests==2.31.0
